+++ R/mt_make_model.R
@@ -12,10 +12,11 @@
 #' 5. Sideways quiet, RAN
 #' 6. Sideways volatile, RAV
 #'
-#'
-#' @details Function is using manually prepared dataset
+#' @details Function is using manually prepared dataset and tries several different random neural network structures.
+#' Once the best neural network is found then the better model is trained and stored.
 #'
 #' @author (C) 2020 Vladimir Zhbanko
+#' @backref Market Type research of Van Tharp Institute: <https://www.vantharp.com/>
 #'
 #' @param num_bars            Number of bars used to detect pattern
 #' @param path_model          Path where the models are be stored
@@ -37,18 +38,18 @@
 #' path_model <- normalizePath(tempdir(),winslash = "/")
 #' path_data <- normalizePath(tempdir(),winslash = "/")
 #'
-#' data(macd_ML2_small)
-#' write_rds(macd_ML2_small, file.path(path_data, 'macd_ML2_small.rds'))
+#' data(macd_ML2)
+#' write_rds(macd_ML2, file.path(path_data, 'macd_ML2.rds'))
 #'
 #' # start h2o engine (using all CPU's by default)
 #' h2o.init()
 #'
 #'
 #' # performing Deep Learning Regression using the custom function
-#' mt_make_model(num_bars = 128,
+#' mt_make_model(num_bars = 64,
 #'               path_model = path_model,
 #'               path_data = path_data,
-#'               f_name_data = "macd_ML2_small.rds")
+#'               f_name_data = "macd_ML2.rds")
 #'
 #' # stop h2o engine
 #' h2o.shutdown(prompt = FALSE)
@@ -70,7 +71,15 @@ mt_make_model <- function(num_bars, path_model, path_data, f_name_data){
 
   macd_ML  <- as.h2o(x = macd_ML2, destination_frame = "macd_ML")
 
-  # fit models from simplest to more complex
+  # try different models and choose the best one...
+  ### random network structure
+  nn_sets <- sample.int(n = 100, 24) %>% matrix(ncol = 3)
+
+  for (i in 1:dim(nn_sets)[1]) {
+
+    # i <- 1
+
+
   ModelC <- h2o.deeplearning(
     model_id = "DL_Classification",
     x = names(macd_ML[,1:num_bars]),
@@ -79,7 +88,7 @@ mt_make_model <- function(num_bars, path_model, path_data, f_name_data){
     activation = "Tanh",
     overwrite_with_best_model = TRUE,
     autoencoder = FALSE,
-    hidden = c(100,100),
+    hidden = nn_sets[i, ],
     loss = "Automatic",
     sparse = TRUE,
     l1 = 1e-4,
@@ -91,6 +100,41 @@ mt_make_model <- function(num_bars, path_model, path_data, f_name_data){
   #ModelC
   #summary(ModelC)
   #h2o.performance(ModelC)
+  RMSE <- h2o::h2o.performance(ModelC)@metrics$RMSE %>% as.data.frame()
+  names(RMSE) <- 'RMSE'
+
+  # record results of modelling
+  if(!exists("df_res")){
+    df_res <- nn_sets[i,] %>% t() %>% as.data.frame() %>% dplyr::bind_cols(RMSE)
+  } else {
+    df_row <- nn_sets[i,] %>% t() %>% as.data.frame() %>% dplyr::bind_cols(RMSE)
+    df_res <- df_res %>% dplyr::bind_rows(df_row)
+  }
+
+
+
+
+  } # end of for loop
+
+  # find which row in the df_res has the smallest RMSE value slice(which.min(Employees))
+  lowest_RMSE <- df_res %>% dplyr::slice(which.min(RMSE)) %>% select(-RMSE) %>% unlist() %>% unname()
+
+  ModelC <- h2o.deeplearning(
+    model_id = "DL_Classification",
+    x = names(macd_ML[,1:num_bars]),
+    y = "M_T",
+    training_frame = macd_ML,
+    activation = "Tanh",
+    overwrite_with_best_model = TRUE,
+    autoencoder = FALSE,
+    hidden = lowest_RMSE,
+    loss = "Automatic",
+    sparse = TRUE,
+    l1 = 1e-4,
+    distribution = "AUTO",
+    stopping_metric = "AUTO",
+    #balance_classes = T,
+    epochs = 200)
 
 h2o.saveModel(ModelC, file.path(path_model, "classification.bin"), force = TRUE)